ai_ocr/pyproject.toml
[tool.poetry]
name = "ai-ocr"
version = "0.1.0"
description = ""
authors = ["Ruo Lee <ruo@illunex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.37.1"
torch = "^2.1.2"
pillow = "^10.2.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
